Config: Quality of life. Variables to enable/disable services.
Based on the number of scales, we can define if we want or not a service to be deployed at the moment.
Bear in mind that only the whoami of traefik is able to deploy more than one service. So I would say that now is only working as a flag to enable/disable.