CA | Traefik | AdGuardHome | Config: Lot of new stuff
Added step-ca to as certificate authority for the intnernal server. The network is aimed to NOT be exposed outside my network, so I don't want to buy any domain or similar. So this is the reason of adding the new service.

Regarding AdGuardHome and the modification of the domain to greenhouse.ogt, is to do it the most generic as possible.

Everything would be customize by the .env file that you create.

Updated the .template.env with the new values.

I ahieve until now the communication between traefik and step-ca. So commiting the achievements.